Update SlashCode.py
Fixed sidebar, fixed general language support, significantly improved file loading system (it uses threading in the background and keeps the GUI, scrolling, etc. smooth, it highlights the document in chunks if the file is large), fixed the minimap and fixed sidebar folder navigation directory trees not working properly and more.