Add new program executable.
This new version of the program executable contains command line files now being loaded properly, even if a previous session file is present. The previous version of Slash Code would properly open and load the command line file being passed as an argument to Slash Code itself, but the session file would be loaded over it. This is now fixed, the session file now gets ignored if a command line file is present as an argument (to sys.argv).