Add Shell Script file for Slash Code.
This Shell Script script is purely for executing Slash Code efficiently on Linux or MacOS instead of manually needing to run the source file of Slash Code. In order to execute the Shell Script file, just open up your dedicated system terminal and write 'chmod +x slashcode.sh', if that works, run this command: './slashcode.sh' (make sure your CWD, the path before the $, is set to Slash Code's directory first, or else this won't work.), if it succeeds, you'll be able to work with Slash Code normally.