Update SlashCode.py
The icons used for the GUI languages now use a more efficient source; PIL (pillow). PIL can convert the images passed onto the TkImage() function to a Tkinter function efficiently, that's why that is used now. And, file saving has been fixed now, it was previously trying to write a variable called 'path', which was undefined and that's the cause of not being able to save the file. This has now been fixed. (Something additional, the GUILANGS dictionary and GUITranslate class have been moved to the beginning of the file to avoid definition errors for functions like show_generror() that uses GUITranslate before the GUITranslate class had been created, now it's ensured that it's defined before anything else.)